+++ src/Console/InitCommand.php
@@ -63,7 +63,7 @@ class InitCommand extends Command
     {
         $this
             ->setName('init')
-            ->setDescription('Initialize deployer system in your project')
+            ->setDescription('Initialize deployer in your project')
             ->addOption('template', 't', InputOption::VALUE_OPTIONAL, 'The template of you project. Available templates: ' . implode(', ', $this->availableTemplates))
             ->addOption('directory', null, InputOption::VALUE_OPTIONAL, 'The directory for create "deploy.php" file', getcwd())
             ->addOption('filename', null, InputOption::VALUE_OPTIONAL, 'The file name. Default "deploy.php"', 'deploy.php');